requirements.txt
pymongo>=4.0
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2
fastapi>=0.95
uvicorn>=0.22
python-dotenv>=1.0
matplotlib>=3.7

